Some compile fixes and new code.
git-svn-id: https://svn.code.sf.net/p/zapping/svn/trunk/vbi@4215 746cafe9-179e-40ba-aef4-f754c8c453d8

+++ contrib/dvbsubs.h
@@ -52,6 +52,6 @@ typedef struct {
 typedef struct {
   char regions[MAX_REGIONS][64];
   int next_region;
-  unsigned int start_pts;
-  unsigned int end_pts;
+  int64_t start_pts;
+  int64_t end_pts;
 } textsub_t;
